feat: 최근 등록된 메모 5개조회

+++ chaekit-spring/src/main/java/qwerty/chaekit/dto/group/activity/ActivityScoreResponse.java
@@ -7,7 +7,7 @@ public record ActivityScoreResponse(
         Long userId,
         String userProfileImageURL,
         String userNickname,
-        int score
+        long score
 ) { 
     public static ActivityScoreResponse of(Long userId, String userProfileImageURL, String userNickname, int score) {
         return ActivityScoreResponse.builder()
